Renaming Pages to (name)-Page Structure
